tech(clean-up): move deprecation and improve date and time picker
1. remove configure screen
2. move InputSingleLine to deprecated
3. move Popover to deprecated
4. Refer UserInput with deprecated Input
5. Deprecate old Tooltip
6. Date and Time picker improvements
