+++ react/src/index.js
@@ -7,12 +7,11 @@ export { default as TwixtDialog } from './Dialog';
 export { default as TwixtDivider } from "./Divider";
 export { default as TwixtHeader } from './Header';
 export { default as TwixtHeading } from './Heading';
-export { default as TwixtSingleLineInput } from './InputSingleLine';
+
 export { default as TwixtLoadingPostCard } from './LoadingPostCard';
 export { default as TwixtModal } from './Modal';
 export { default as TwixtDropdown } from './OneSelectDropdown';
 export { default as TwixtPageProgressLine } from './PageProgressLine';
-export { default as TwixtPopover } from './Popover';
 export { default as TwixtPostItemList } from './PostItemList';
 export { default as TwixtSimpleCodeEditor } from './SimpleCodeEditor';
 export { default as SimpleTable } from './SimpleTable';
@@ -56,3 +55,5 @@ export { default as TwixtInputSingleLine } from './Inputs/InputSingleLine';
 export { default as TwixtButton_DEPRECATED } from './_deprecated/Button';
 export { default as TwixtToolTip_DEPRECATED } from './_deprecated/Tooltip';
 export { default as TwixtCheckbox_DEPRECATED } from './_deprecated/Checkbox';
+export { default as TwixtSingleLineInput_DEPRECATED } from './InputSingleLine';
+export { default as TwixtPopover_DEPRECATED } from './Popover';
